+++ crates/rattler-bin/src/commands/create.rs
@@ -67,8 +67,13 @@ pub struct Opt {
     timeout: Option<u64>,
 
     /// Target prefix (environment path) for package installation
-    #[clap(short = 'p', long = "prefix", visible_alias = "target-prefix")]
-    target_prefix: Option<PathBuf>,
+    #[clap(
+        short = 'p',
+        long = "prefix",
+        visible_alias = "target-prefix",
+        default_value = ".prefix"
+    )]
+    target_prefix: PathBuf,
 
     #[clap(long)]
     strategy: Option<SolveStrategy>,
@@ -122,14 +127,8 @@ impl From<SolveStrategy> for rattler_solve::SolveStrategy {
 pub async fn create(opt: Opt) -> miette::Result<()> {
     let channel_config =
         ChannelConfig::default_with_root_dir(env::current_dir().into_diagnostic()?);
-    let current_dir = env::current_dir().into_diagnostic()?;
-    let target_prefix = opt
-        .target_prefix
-        .unwrap_or_else(|| current_dir.join(".prefix"));
-
     // Make the target prefix absolute
-    let target_prefix = std::path::absolute(target_prefix).into_diagnostic()?;
-    println!("Target prefix: {}", target_prefix.display());
+    let target_prefix = std::path::absolute(opt.target_prefix).into_diagnostic()?;
 
     // Determine the platform we're going to install for
     let install_platform = if let Some(platform) = opt.platform {
@@ -243,7 +242,7 @@ pub async fn create(opt: Opt) -> miette::Result<()> {
                 .collect::<miette::Result<Vec<_>>>()?)
         } else {
             rattler_virtual_packages::VirtualPackage::detect(
-                &rattler_virtual_packages::VirtualPackageOverrides::default(),
+                &rattler_virtual_packages::VirtualPackageOverrides::from_env(),
             )
             .map(|vpkgs| {
                 vpkgs